Fixed issues with my code; I had completely misunderstood several key concepts on this assignment, I've gone over the material again, and solved the issues that were found lacking. -I adjusted the LogFactory class to use the appropriate initializer for the ClassName. -Removed the hard coding within the CreateLogger method. -Changed filepath to private member. -CreateLogger will  now return null at the appropriate times. -Fixed my copy paste issues within the BaseLoggerMixins class. -Use the nameof() in the logger to identify the class now.

+++ Logger/FileLogger.cs
@@ -7,31 +7,33 @@ namespace Logger
 {
     public class FileLogger : BaseLogger
     {
-        public string FilePath { get; set; }
-        public override string ClassName { get; set; }
+        private string? filePath;
+        public string FilePath
+        {
+            get => filePath!;
+            set => filePath = value ?? throw new ArgumentNullException(null);
+        }
+
         public FileLogger(string filePath)
         {
             FilePath = filePath;
-            ClassName = "FileLogger";
         }
+
         public override void Log(LogLevel logLevel, string message)
         {
-            StreamWriter appendMessage;
-            try
+            TextWriter appendMessage;
+            if (FilePath != null && File.Exists(FilePath))
             {
                 appendMessage = new StreamWriter(FilePath);
+                string dateTime = DateTime.Now.ToString("yyyy-MM-dd_HH:mm");
+                appendMessage.WriteLine("Date/time: " + dateTime);
+                appendMessage.WriteLine(base.ClassName);
+                appendMessage.WriteLine(logLevel);
+                appendMessage.WriteLine(message);
+                appendMessage.Close();
+            }else{
+                throw new FileNotFoundException();
             }
-            catch (ArgumentNullException)
-            {
-                throw new ArgumentNullException(FilePath);
-            }
-
-            string dateTime = DateTime.Now.ToString("yyyy-MM-dd_HH:mm");
-            appendMessage.WriteLine("Date/time: " + dateTime);
-            appendMessage.WriteLine(ClassName);
-            appendMessage.WriteLine(logLevel);
-            appendMessage.WriteLine(message);
-            appendMessage.Dispose();
         }
     }
 }
